switched to big decimal

+++ src/main/java/org/weather/app/dto/CityDto.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -28,9 +29,9 @@ public class CityDto {
     @Setter
     public static class CoordDto {
         @JsonProperty("lon")
-        private Double longitude;
+        private BigDecimal longitude;
 
         @JsonProperty("lat")
-        private Double latitude;
+        private BigDecimal latitude;
     }
 }
